attempt to generalize state handling.
This strategy seems like it might not be quite right, which is why I put it in a separate branch. Having some kind of generic solution to eliminate shared code is a good idea, but templating is probably going to lead to more headaches and some amount of duplication in the long run.

Might be better to do a take-two where we have a shared 'actor' parent that has some fundamental stuff all of these things - enemies and player - have in common.